# Server/src/configs/api.yaml
openapi: 3.0.3
info:
  title: Swagger for ecommerce - OpenAPI 3.0
  description: Mô tả API của SunCinema for frontend
  termsOfService: http://swagger.io/terms/
  license:
    name: Github
    url: https://github.com/duon97ran2/SunCinema
  version: 1.0.11
servers:
  - url: http://localhost:5000/api/

# Tags
...
paths:
...
  /food:
    get:
      tags:
        - food
      summary: Lấy ra tất cả đồ ăn
      responses:
        "200":
          description: Lấy đồ ăn thành công!
        "400":
          description: Không lấy được đồ ăn!
    post:
      tags:
        - food
      summary: Thêm đồ ăn
      description: Thêm đồ ăn
      requestBody:
        description: Bắt buộc phải có các trường giống bên dưới
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Food"
        required: true
      responses:
        "200":
          description: Thêm đồ ăn thành công!
        "400":
          description: Không thêm được đồ ăn!
...
components:
  schemas:
...
    Food:
      properties:
        name:
          type: string
        price:
          type: number
        stock:
          type: string
        size:
          type: number
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT